backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyotp
pytest
httpx
psycopg2-binary
python-dotenv
reportlab
alembic
gunicorn
